Address review and fix Windows builds
