+++ Modules/_testcapimodule.c
@@ -5754,7 +5754,10 @@ PyInit__testcapi(void)
     if (_PyTestCapi_Init_GetArgs(m) < 0) {
         return NULL;
     }
-    if (_PyTestCapi_Init_Time(m) < 0) {
+    if (_PyTestCapi_Init_PyTime(m) < 0) {
+        return NULL;
+    }
+    if (_PyTestCapi_Init_DateTime(m) < 0) {
         return NULL;
     }
